+++ resources/lang/de/app.php
@@ -249,6 +249,7 @@ return [
     'search' => 'Suchen',
     'search_ellipsis' => 'Suchen...',
     'type_to_search' => 'Zum Suchen tippen...',
+    'type_to_search_fields' => 'Zum Suchen tippen (:fields)...',
     'not_all_results_listed_please_refine_criteria' => 'Nich alle Resultate können angezeigt werden. Bitte Suchkriterien verfeinern.',
     'no_results_found_for_term' => 'Keine Resultate gefunden für \':term\'.',
     'male' => 'männlich',